fix(codex): harden Windows package layout repair
